+++ Packages/Sources/TiledownCLI/Entrypoint.swift
@@ -1,9 +1,18 @@
+import Foundation
+
 @main
 struct Entrypoint {
-    static func main() throws {
-        try Command(
-            arguments: Array(CommandLine.arguments.dropFirst()),
-        )
-        .run()
+    static func main() {
+        do {
+            try Command(
+                arguments: Array(CommandLine.arguments.dropFirst()),
+            )
+            .run()
+        } catch {
+            FileHandle.standardError.write(
+                Data((String(describing: error) + "\n").utf8),
+            )
+            exit(1)
+        }
     }
 }